src/CatalogAPI/Services/ProductService.cs
namespace CatalogAPI.Services
{
    public class ProductService : IProductService
    {
        private IProductRepository _productRepository;

        public ProductService(IProductRepository productRepository)
        {
            _productRepository = productRepository;
        }
        public IEnumerable<Product> GetAllProducts()
        {
            return _productRepository.GetAllProducts();
        }

        public Product GetProductById(long productId)
        {
            return _productRepository.GetProductById(productId);
        }
    }
}